+++ R/window.R
@@ -10,16 +10,18 @@
 #' @param .f A function or one-sided formula using purrr-like syntax. If a 
 #' formula, it is converted to a function.
 #' @param ... Additional arguments passed on to `.f`.
-#' @param size An integer for window size.
-#' @param fill A single value or data frame to replace `NA`.
+#' @param .size An integer for window size.
+#' @param .fill A single value or data frame to replace `NA`.
 #' @param .id If not `NULL` a variable with this name will be created giving 
 #' either the name or the index of the data frame, which is passed to 
 #' [dplyr::bind_rows].
 #'
 #' @rdname slide
 #' @export
-#' @seealso [tile] for tiling window without overlapping observations;
-#' [stretch] for expanding more observations
+#' @seealso 
+#' * [slide2], [pslide]
+#' * [tile] for tiling window without overlapping observations
+#' * [stretch] for expanding more observations
 #' @details The `slide()` function attempts to tackle more general problems using
 #' the purrr-like syntax. For some specialist functions like `mean` and `sum`, 
 #' you may like to check out for 
@@ -28,9 +30,9 @@
 #' @examples
 #' # sliding through a vector ----
 #' x <- 1:10
-#' slide(x, mean, size = 3)
-#' slide(x, ~ mean(.), size = 3)
-#' slide(x, mean, size = 3, fill = 0)
+#' slide(x, mean, .size = 3)
+#' slide(x, ~ mean(.), .size = 3)
+#' slide(x, mean, .size = 3, .fill = 0)
 #'
 #' \dontrun{
 #' # takes a little longer for cran check
@@ -45,118 +47,122 @@
 #'     ~ slide_dfr(., function(x) {
 #'       fit <- lm(Count ~ Time, data = x)
 #'       data.frame(fitted = fitted(fit), resid = residuals(fit))
-#'     }, size = 48)
+#'     }, .size = 48)
 #'   )
 #' diag_jan[48:57, ]
 #' # save lm models as additional columns
 #' lm_jan <- jan %>% 
 #'   purrr::map(
-#'     ~ mutate(., lm = slide_lst(., ~ lm(Count ~ Time, data = .), size = 48)
+#'     ~ mutate(., lm = slide_lst(., ~ lm(Count ~ Time, data = .), .size = 48)
 #'   ))
 #' lm_jan[[1]][48:57, ]
 #' }
 #' @export
-slide <- function(.x, .f, ..., size = 1, fill = NA_real_) {
-  lst_x <- slider(.x, size = size)
-  c(rep_len(fill, size - 1), purrr::map_dbl(lst_x, .f, ...))
+slide <- function(.x, .f, ..., .size = 1, .fill = NA_real_) {
+  lst_x <- slider(.x, .size = .size)
+  c(rep_len(.fill, .size - 1), purrr::map_dbl(lst_x, .f, ...))
 }
 
 #' @rdname slide
 #' @export
-slide_lst <- function(.x, .f, ..., size = 1, fill = NA) {
-  lst_x <- slider(.x, size = size)
+slide_lst <- function(.x, .f, ..., .size = 1, .fill = NA) {
+  lst_x <- slider(.x, .size = .size)
   result <- purrr::map(lst_x, .f, ...)
-  if (is.na(fill)) {
-    fill <- rep_len(fill, length(result[[1]]))
+  if (is.na(.fill)) {
+    .fill <- rep_len(.fill, length(result[[1]]))
   }
-  c(replicate(n = size - 1, fill, simplify = FALSE), result)
+  c(replicate(n = .size - 1, .fill, simplify = FALSE), result)
 }
 
 #' @rdname slide
 #' @export
-slide_dfr <- function(.x, .f, ..., size = 1, fill = NA, .id = NULL) {
-  out <- slide_lst(.x, .f = .f, ..., size = size, fill = fill)
-  out_named <- purrr::map(out, `names<-`, names(out[[size]]))
+slide_dfr <- function(.x, .f, ..., .size = 1, .fill = NA, .id = NULL) {
+  out <- slide_lst(.x, .f = .f, ..., .size = .size, .fill = .fill)
+  out_named <- purrr::map(out, `names<-`, names(out[[.size]]))
   dplyr::bind_rows(!!! out_named, .id = .id)
 }
 
 # #' @rdname slide
 # #' @export
-# slide_dfc <- function(x, .f, ..., size = 1, fill = NA) {
-#   out <- slide_lst(x, .f = .f, ..., size = size, fill = fill)
-#   out_named <- purrr::map(out, `names<-`, names(out[[size]]))
+# slide_dfc <- function(x, .f, ..., .size = 1, .fill = NA) {
+#   out <- slide_lst(x, .f = .f, ..., .size = .size, .fill = .fill)
+#   out_named <- purrr::map(out, `names<-`, names(out[[.size]]))
 #   dplyr::bind_cols(!!! out_named)
 # }
 
 #' Sliding window calculation over multiple inputs simultaneously
 #'
-#' @param .x,.y A vector of numerics, or data frame. If a data frame, row-wise rolling
-#' window is performed.
+#' @param .x,.y A vector of numerics, or data frame. If a data frame, row-wise 
+#' rolling window is performed.
 #' @inheritParams slide
 #'
 #' @rdname slide2
 #' @export
-#' @seealso [tile2] for tiling window without overlapping observations;
-#' [stretch2] for expanding more observations
+#' @seealso 
+#' * [slide]
+#' * [tile2] for tiling window without overlapping observations
+#' * [stretch2] for expanding more observations
 #'
+#' @export
 #' @examples
 #' x <- 1:10
 #' y <- 10:1
-#' slide2(x, y, cor, size = 3)
-#' slide2_lst(x, y, cor, size = 3)
-#' @export
-slide2 <- function(.x, .y, .f, ..., size = 1, fill = NA_real_) {
-  lst <- slider(.x, .y, size = size)
-  c(rep_len(fill, size - 1), purrr::map2_dbl(lst[[1]], lst[[2]], .f, ...))
+#' z <- x * 2
+#' slide2(x, y, cor, .size = 3)
+#' slide2_lst(x, y, cor, .size = 3)
+#' pslide(list(x, y, z), sum, .size = 3)
+slide2 <- function(.x, .y, .f, ..., .size = 1, .fill = NA_real_) {
+  lst <- slider(.x, .y, .size = .size)
+  c(rep_len(.fill, .size - 1), purrr::map2_dbl(lst[[1]], lst[[2]], .f, ...))
 }
 
 #' @rdname slide2
 #' @export
-slide2_lst <- function(.x, .y, .f, ..., size = 1, fill = NA) {
-  lst <- slider(.x, .y, size = size)
+slide2_lst <- function(.x, .y, .f, ..., .size = 1, .fill = NA) {
+  lst <- slider(.x, .y, .size = .size)
   result <- purrr::map2(lst[[1]], lst[[2]], .f, ...)
-  if (is.na(fill)) {
-    fill <- rep_len(fill, length(result[[1]]))
+  if (is.na(.fill)) {
+    .fill <- rep_len(.fill, length(result[[1]]))
   }
-  c(replicate(n = size - 1, fill, simplify = FALSE), result)
+  c(replicate(n = .size - 1, .fill, simplify = FALSE), result)
 }
 
 #' @rdname slide2
 #' @export
-slide2_dfr <- function(.x, .y, .f, ..., size = 1, fill = NA, .id = NULL) {
-  out <- slide_lst(.x, .y, .f = .f, ..., size = size, fill = fill)
-  out_named <- purrr::map(out, `names<-`, names(out[[size]]))
+slide2_dfr <- function(.x, .y, .f, ..., .size = 1, .fill = NA, .id = NULL) {
+  out <- slide_lst(.x, .y, .f = .f, ..., .size = .size, .fill = .fill)
+  out_named <- purrr::map(out, `names<-`, names(out[[.size]]))
   dplyr::bind_rows(!!! out_named, .id = .id)
 }
 
 #' @rdname slide2
 #' @inheritParams purrr::pmap
 #' @export
-pslide <- function(.l, .f, ..., size = 1, fill = NA_real_) {
-  lst <- slider(.l, size = size)
-  c(rep_len(fill, size - 1), purrr::pmap_dbl(lst, .f, ...))
+pslide <- function(.l, .f, ..., .size = 1, .fill = NA_real_) {
+  lst <- slider(.l, .size = .size)
+  c(rep_len(.fill, .size - 1), purrr::pmap_dbl(lst, .f, ...))
 }
 
 #' @rdname slide2
 #' @export
-pslide_lst <- function(.l, .f, ..., size = 1, fill = NA) {
-  lst <- slider(.l, size = size)
+pslide_lst <- function(.l, .f, ..., .size = 1, .fill = NA) {
+  lst <- slider(.l, .size = .size)
   result <- purrr::pmap(lst, .f, ...)
-  if (is.na(fill)) {
-    fill <- rep_len(fill, length(result[[1]]))
+  if (is.na(.fill)) {
+    .fill <- rep_len(.fill, length(result[[1]]))
   }
-  c(replicate(n = size - 1, fill, simplify = FALSE), result)
+  c(replicate(n = .size - 1, .fill, simplify = FALSE), result)
 }
 
 #' @rdname slide2
 #' @export
-pslide_dfr <- function(.l, .f, ..., size = 1, fill = NA, .id = NULL) {
-  out <- pslide_lst(.l, .f = .f, ..., size = size, fill = fill)
-  out_named <- purrr::pmap(out, `names<-`, names(out[[size]]))
+pslide_dfr <- function(.l, .f, ..., .size = 1, .fill = NA, .id = NULL) {
+  out <- pslide_lst(.l, .f = .f, ..., .size = .size, .fill = .fill)
+  out_named <- purrr::pmap(out, `names<-`, names(out[[.size]]))
   dplyr::bind_rows(!!! out_named, .id = .id)
 }
 
-#' Splits the input to a list according to the rolling window size.
+#' Splits the input to a list according to the rolling window .size.
 #' 
 #' @param ... Objects to be splitted.
 #' @inheritParams slide
@@ -164,26 +170,26 @@ pslide_dfr <- function(.l, .f, ..., size = 1, fill = NA, .id = NULL) {
 #' @export
 #' @examples
 #' x <- 1:10
-#' slider(x, size = 3)
+#' slider(x, .size = 3)
 #' y <- 10:1
-#' slider(x, y, size = 3)
-slider <- function(..., size = 1) {
+#' slider(x, y, .size = 3)
+slider <- function(..., .size = 1) {
   x <- flatten(list(...))
   if (has_length(x, 1)) {
-    return(slider_base(x[[1]], size = size))
+    return(slider_base(x[[1]], .size = .size))
   } else {
-    return(purrr::map(x, ~ slider_base(., size = size)))
+    return(purrr::map(x, ~ slider_base(., .size = .size)))
   }
 }
 
-slider_base <- function(x, size = 1) {
-  bad_window_function(x, size)
+slider_base <- function(x, .size = 1) {
+  bad_window_function(x, .size)
   len_x <- NROW(x)
-  lst_idx <- seq_len(len_x - size + 1)
+  lst_idx <- seq_len(len_x - .size + 1)
   if (is_atomic(x)) {
-    return(purrr::map(lst_idx, ~ x[(.):(. + size - 1)]))
+    return(purrr::map(lst_idx, ~ x[(.):(. + .size - 1)]))
   }
-  purrr::map(lst_idx, ~ x[(.):(. + size - 1), , drop = FALSE])
+  purrr::map(lst_idx, ~ x[(.):(. + .size - 1), , drop = FALSE])
 }
 
 #' Tiling window calculation
@@ -197,37 +203,39 @@ slider_base <- function(x, size = 1) {
 #'
 #' @rdname tile
 #' @export
-#' @seealso [slide] for sliding window with overlapping observations;
-#' [stretch] for expanding more observations
+#' @seealso 
+#' * [tile2], [ptile]
+#' * [slide] for sliding window with overlapping observations
+#' * [stretch] for expanding more observations
 #'
 #' @examples
 #' # tiling over a vector ----
 #' x <- 1:10
-#' tile(x, sum, size = 3)
-#' tile(x, ~ sum(.), size = 3)
+#' tile(x, sum, .size = 3)
+#' tile(x, ~ sum(.), .size = 3)
 #'
 #' # tiling over a 2-day window for hourly data ----
 #' \dontrun{
 #' pedestrian %>% 
 #'   split_by(Sensor) %>% 
-#'   purrr::map_dfr(~ tile_dfr(., ~ quantile(.$Count), size = 48))
+#'   purrr::map_dfr(~ tile_dfr(., ~ quantile(.$Count), .size = 48))
 #' }
-tile <- function(.x, .f, ..., size = 1) {
-  lst_x <- tiler(.x, size = size)
+tile <- function(.x, .f, ..., .size = 1) {
+  lst_x <- tiler(.x, .size = .size)
   purrr::map_dbl(lst_x, .f, ...)
 }
 
 #' @rdname tile
 #' @export
-tile_lst <- function(.x, .f, ..., size = 1) {
-  lst_x <- tiler(.x, size = size)
+tile_lst <- function(.x, .f, ..., .size = 1) {
+  lst_x <- tiler(.x, .size = .size)
   purrr::map(lst_x, .f, ...)
 }
 
 #' @rdname tile
 #' @export
-tile_dfr <- function(.x, .f, ..., size = 1, .id = NULL) {
-  out <- tile_lst(.x = .x, .f = .f, ..., size = size)
+tile_dfr <- function(.x, .f, ..., .size = 1, .id = NULL) {
+  out <- tile_lst(.x = .x, .f = .f, ..., .size = .size)
   dplyr::bind_rows(!!! out, .id = .id)
 }
 
@@ -240,58 +248,62 @@ tile_dfr <- function(.x, .f, ..., size = 1, .id = NULL) {
 #'
 #' @rdname tile2
 #' @export
-#' @seealso [slide2] for sliding window with overlapping observations;
-#' [stretch2] for expanding more observations
+#' @seealso 
+#' * [tile]
+#' * [slide2] for sliding window with overlapping observations
+#' * [stretch2] for expanding more observations
 #'
 #' @examples
 #' x <- 1:10
 #' y <- 10:1
-#' tile2(x, y, cor, size = 5)
-#' tile2_lst(x, y, ~ cor(.x, .y), size = 5)
-tile2 <- function(.x, .y, .f, ..., size = 1) {
-  lst <- tiler(.x, .y, size = size)
+#' z <- x * 2
+#' tile2(x, y, cor, .size = 5)
+#' tile2_lst(x, y, ~ cor(.x, .y), .size = 5)
+#' ptile(list(x, y, z), sum, .size = 5)
+tile2 <- function(.x, .y, .f, ..., .size = 1) {
+  lst <- tiler(.x, .y, .size = .size)
   purrr::map2_dbl(lst[[1]], lst[[2]], .f, ...)
 }
 
 #' @rdname tile2
 #' @export
-tile2_lst <- function(.x, .y, .f, ..., size = 1) {
-  lst <- tiler(.x, .y, size = size)
+tile2_lst <- function(.x, .y, .f, ..., .size = 1) {
+  lst <- tiler(.x, .y, .size = .size)
   purrr::map2(lst[[1]], lst[[2]], .f, ...)
 }
 
 #' @rdname tile2
 #' @export
-tile2_dfr <- function(.x, .y, .f, ..., size = 1, .id = NULL) {
-  out <- tile2_lst(.x, .y, .f = .f, ..., size = size)
+tile2_dfr <- function(.x, .y, .f, ..., .size = 1, .id = NULL) {
+  out <- tile2_lst(.x, .y, .f = .f, ..., .size = .size)
   dplyr::bind_rows(!!! out, .id = .id)
 }
 
 #' @rdname tile2
 #' @export
-ptile <- function(.l, .f, ..., size = 1) {
-  lst <- tiler(.l, size = size)
+ptile <- function(.l, .f, ..., .size = 1) {
+  lst <- tiler(.l, .size = .size)
   purrr::pmap_dbl(lst, .f, ...)
 }
 
 #' @rdname tile2
 #' @export
-ptile_lst <- function(.l, .f, ..., size = 1) {
-  lst <- tiler(.l, size = size)
+ptile_lst <- function(.l, .f, ..., .size = 1) {
+  lst <- tiler(.l, .size = .size)
   purrr::pmap(lst, .f, ...)
 }
 
 #' @rdname tile2
 #' @export
-ptile_dfr <- function(.l, .f, ..., size = 1, .id = NULL) {
-  out <- tile2_lst(.l, .f = .f, ..., size = size)
+ptile_dfr <- function(.l, .f, ..., .size = 1, .id = NULL) {
+  out <- tile2_lst(.l, .f = .f, ..., .size = .size)
   dplyr::bind_rows(!!! out, .id = .id)
 }
 
 # #' @rdname tile
 # #' @export
-# tile_dfc <- function(x, .f, ..., size = 1) {
-#   out <- tile_lst(x = x, .f = .f, ..., size = size)
+# tile_dfc <- function(x, .f, ..., .size = 1) {
+#   out <- tile_lst(x = x, .f = .f, ..., .size = .size)
 #   dplyr::bind_cols(!!! out)
 # }
 
@@ -299,24 +311,24 @@ ptile_dfr <- function(.l, .f, ..., size = 1, .id = NULL) {
 #' @export
 #' @examples
 #' x <- 1:10
-#' tiler(x, size = 3)
+#' tiler(x, .size = 3)
 #' y <- 10:1
-#' tiler(x, y, size = 3)
-tiler <- function(..., size = 1) {
+#' tiler(x, y, .size = 3)
+tiler <- function(..., .size = 1) {
   x <- flatten(list(...))
   if (has_length(x, 1)) {
-    return(tiler_base(x[[1]], size = size))
+    return(tiler_base(x[[1]], .size = .size))
   } else {
-    return(purrr::map(x, ~ tiler_base(., size = size)))
+    return(purrr::map(x, ~ tiler_base(., .size = .size)))
   }
 }
 
-tiler_base <- function(x, size = 1) {
-  bad_window_function(x, size)
+tiler_base <- function(x, .size = 1) {
+  bad_window_function(x, .size)
   len_x <- NROW(x)
   seq_x <- seq_len(len_x)
   denom <- len_x + 1
-  frac <- ceiling((seq_x %% denom) / size)
+  frac <- ceiling((seq_x %% denom) / .size)
   unname(split(x, frac, drop = TRUE))
 }
 
@@ -328,130 +340,137 @@ tiler_base <- function(x, size = 1) {
 #' * `stretch_dfr()` return data frame using row-binding
 #'
 #' @inheritParams slide
-#' @param size,init An integer for moving and initial window size.
+#' @param .size,.init An integer for moving and initial window size.
 #'
 #' @rdname stretch
 #' @export
-#' @seealso [slide] for sliding window with overlapping observations;
-#' [tile] for tiling window without overlapping observations.
+#' @seealso 
+#' * [stretch2], [pstretch]
+#' * [slide] for sliding window with overlapping observations
+#' * [tile] for tiling window without overlapping observations
 #'
 #' @examples
 #' x <- 1:10
-#' stretch(x, mean, init = 3)
-#' stretch(x, ~ mean(.), init = 3)
+#' stretch(x, mean, .init = 3)
+#' stretch(x, ~ mean(.), .init = 3)
 #'
 #' # stretching a 2-day window for a data frame ----
 #' sx <- pedestrian %>% 
 #'   filter(Sensor == "Southern Cross Station", Date <= as.Date("2015-01-10"))
 #' sx %>%
-#'   stretch_dfr(~ quantile(.$Count), init = 48)
-stretch <- function(.x, .f, ..., size = 1, init = 1) {
-  lst_x <- stretcher(.x, size = size, init = init)
+#'   stretch_dfr(~ quantile(.$Count), .init = 48)
+stretch <- function(.x, .f, ..., .size = 1, .init = 1) {
+  lst_x <- stretcher(.x, .size = .size, .init = .init)
   purrr::map_dbl(lst_x, .f, ...)
 }
 
 #' @rdname stretch
 #' @export
-stretch_lst <- function(.x, .f, ..., size = 1, init = 1) {
-  lst_x <- stretcher(.x, size = size, init = init)
+stretch_lst <- function(.x, .f, ..., .size = 1, .init = 1) {
+  lst_x <- stretcher(.x, .size = .size, .init = .init)
   purrr::map(lst_x, .f, ...)
 }
 
 #' @rdname stretch
 #' @export
-stretch_dfr <- function(.x, .f, ..., size = 1, init = 1, .id = NULL) {
-  out <- stretch_lst(.x, .f = .f, ..., size = size, init = init)
+stretch_dfr <- function(.x, .f, ..., .size = 1, .init = 1, .id = NULL) {
+  out <- stretch_lst(.x, .f = .f, ..., .size = .size, .init = .init)
   dplyr::bind_rows(!!! out, .id = .id)
 }
 
 # #' @rdname stretch
 # #' @export
-# stretch_dfc <- function(x, .f, ..., size = 1, init = 1) {
-#   out <- stretch_lst(x, .f = .f, ..., size = size, init = init)
+# stretch_dfc <- function(x, .f, ..., .size = 1, .init = 1) {
+#   out <- stretch_lst(x, .f = .f, ..., .size = .size, .init = .init)
 #   dplyr::bind_cols(!!! out)
 # }
 
 #' Stretching window calculation over multiple simultaneously
 #'
 #' @inheritParams slide2
-#' @param size,init An integer for moving and initial window size.
+#' @param .size,.init An integer for moving and initial window size.
 #'
 #' @rdname stretch2
 #' @export
-#' @seealso [slide2] for sliding window with overlapping observations;
-#' [tile2] for tiling window without overlapping observations.
+#' @seealso 
+#' * [stretch]
+#' * [slide2] for sliding window with overlapping observations
+#' * [tile2] for tiling window without overlapping observations
 #'
 #' @examples
 #' x <- 1:10
 #' y <- 10:1
-#' stretch2(x, y, cor, init = 3)
-#' stretch2(x, y, ~ cor(.x, .y), init = 3)
-stretch2 <- function(.x, .y, .f, ..., size = 1, init = 1) {
-  lst <- stretcher(.x, .y, size = size, init = init)
+#' z <- x * 2
+#' stretch2(x, y, cor, .init = 3)
+#' stretch2(x, y, ~ cor(.x, .y), .init = 3)
+#' pstretch(list(x, y, z), sum, .init = 3)
+stretch2 <- function(.x, .y, .f, ..., .size = 1, .init = 1) {
+  lst <- stretcher(.x, .y, .size = .size, .init = .init)
   purrr::map2_dbl(lst[[1]], lst[[2]], .f, ...)
 }
 
 #' @rdname stretch2
 #' @export
-stretch2_lst <- function(.x, .y, .f, ..., size = 1, init = 1) {
-  lst <- stretcher(.x, .y, size = size, init = init)
+stretch2_lst <- function(.x, .y, .f, ..., .size = 1, .init = 1) {
+  lst <- stretcher(.x, .y, .size = .size, .init = .init)
   purrr::map2(lst[[1]], lst[[2]], .f, ...)
 }
 
 #' @rdname stretch2
 #' @export
-stretch2_dfr <- function(.x, .y, .f, ..., size = 1, init = 1, .id = NULL) {
-  out <- stretch_lst(.x, .y, .f = .f, ..., size = size, init = init)
+stretch2_dfr <- function(.x, .y, .f, ..., .size = 1, .init = 1, .id = NULL) {
+  out <- stretch_lst(.x, .y, .f = .f, ..., .size = .size, .init = .init)
   dplyr::bind_rows(!!! out, .id = .id)
 }
 
 #' @rdname stretch2
 #' @export
-pstretch <- function(.l, .f, ..., size = 1, init = 1) {
-  lst <- stretcher(.l, size = size, init = init)
+pstretch <- function(.l, .f, ..., .size = 1, .init = 1) {
+  lst <- stretcher(.l, .size = .size, .init = .init)
   purrr::pmap_dbl(.l, .f, ...)
 }
 
 #' @rdname stretch2
 #' @export
-pstretch_lst <- function(.l, .f, ..., size = 1, init = 1) {
-  lst <- stretcher(.l, size = size, init = init)
+pstretch_lst <- function(.l, .f, ..., .size = 1, .init = 1) {
+  lst <- stretcher(.l, .size = .size, .init = .init)
   purrr::pmap(.l, .f, ...)
 }
 
 #' @rdname stretch2
 #' @export
-pstretch_dfr <- function(.l, .f, ..., size = 1, init = 1, .id = NULL) {
-  out <- stretch_lst(.l, .f = .f, ..., size = size, init = init)
+pstretch_dfr <- function(.l, .f, ..., .size = 1, .init = 1, .id = NULL) {
+  out <- stretch_lst(.l, .f = .f, ..., .size = .size, .init = .init)
   dplyr::bind_rows(!!! out, .id = .id)
 }
 
 #' @rdname slider
+#' @param .size,.init An integer for moving and initial window size.
 #' @export
 #' @examples
 #' x <- 1:10
-#' stretcher(x, init = 3)
+#' stretcher(x, .init = 3)
 #' y <- 10:1
-#' stretcher(x, y, init = 3)
-stretcher <- function(..., size = 1, init = 1) {
+#' stretcher(x, y, .init = 3)
+stretcher <- function(..., .size = 1, .init = 1) {
   x <- flatten(list(...))
   if (has_length(x, 1)) {
-    return(stretcher_base(x[[1]], size = size, init = init))
+    return(stretcher_base(x[[1]], .size = .size, .init = .init))
   } else {
-    return(purrr::map(x, ~ stretcher_base(., size = size, init = init)))
+    return(purrr::map(x, ~ stretcher_base(., .size = .size, .init = .init)))
   }
 }
 
-stretcher_base <- function(x, size = 1, init = 1) {
-  bad_window_function(x, size)
-  if (!is_bare_numeric(init, n = 1) || init < 1) {
-    abort("`init` must be a positive integer.")
+stretcher_base <- function(x, .size = 1, .init = 1) {
+  bad_window_function(x, .size)
+  if (!is_bare_numeric(.init, n = 1) || .init < 1) {
+    abort("`.init` must be a positive integer.")
   }
-  counter <- incr(init = init, size = size) 
+  counter <- incr(init = .init, size = .size) 
 
-  ncall <- seq_len(ceiling((NROW(x) - init) / size) - 1)
+  ncall <- seq_len(ceiling((NROW(x) - .init) / .size) - 1)
   incr_lst <- c(
-    list(seq_len(init)), 
+    list(seq_len(.init)), 
     purrr::map(ncall, ~ seq_len(counter())),
     list(seq_along(x))
   )
@@ -461,12 +480,12 @@ stretcher_base <- function(x, size = 1, init = 1) {
   purrr::map(incr_lst, ~ x[., , drop = FALSE])
 }
 
-bad_window_function <- function(.x, size) {
+bad_window_function <- function(.x, .size) {
   if (is_bare_list(.x)) {
     abort("`.x` must not be a list of lists.")
   }
-  if (!is_bare_numeric(size, n = 1) || size < 1) {
-    abort("`size` must be a positive integer.")
+  if (!is_bare_numeric(.size, n = 1) || .size < 1) {
+    abort("`.size` must be a positive integer.")
   }
 }